restructure benches directory/modules
