+++ libursa/Cargo.toml
@@ -8,7 +8,6 @@ readme =  "../README.md"
 repository = "https://github.com/hyperledger/ursa"
 documentation = "https://docs.rs/ursa"
 homepage = "https://crates.io/crates/ursa"
-autobenches = false
 
 [lib]
 name = "ursa"
